requirements.txt
bitsandbytes==0.48.2
datasets==4.0.0
gensim==4.4.0
numpy==1.26.4
pandas==2.2.3
peft==0.18.0
PyYAML==6.0.2
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
tokenizers==0.21.4
torch==2.7.1+cu118
transformers==4.55.3

